+++ Tests/ParseSwiftTests/ParseAnalyticsTests.swift
@@ -1,5 +1,5 @@
 //
-//  ParseAnalytics.swift
+//  ParseAnalyticsTests.swift
 //  ParseSwift
 //
 //  Created by Corey Baker on 5/20/21.
@@ -34,49 +34,36 @@ class ParseAnalyticsTests: XCTestCase {
         try ParseStorage.shared.deleteAll()
     }
 
-    func testCheckCommand() throws {
+    func testSaveCommand() throws {
         let name = "hello"
         let event = ParseAnalytics(name: name)
         let command = event.saveCommand()
         XCTAssertEqual(command.path.urlComponent, "/events/\(name)")
         XCTAssertEqual(command.method, API.Method.POST)
-        XCTAssertNil(command.body)
-    
+        XCTAssertNotNil(command.body)
+        XCTAssertNil(command.body?.at)
+        XCTAssertNil(command.body?.dimensions)
+
         let date = Date()
-        let event2 = ParseAnalytics(name: name, at: date, dimensions: ["stop" : "drop"])
+        let dimensions = ["stop": "drop"]
+        var event2 = ParseAnalytics(name: name, dimensions: dimensions, at: date)
         let command2 = event2.saveCommand()
         XCTAssertEqual(command2.path.urlComponent, "/events/\(name)")
         XCTAssertEqual(command2.method, API.Method.POST)
-        XCTAssertNil(command2.body)
+        XCTAssertNotNil(command2.body)
+        XCTAssertEqual(command2.body?.at, date)
+        XCTAssertEqual(command2.body?.dimensions, dimensions)
+
+        event2.at = nil //Clear date for comparison
+        let encoded = try ParseCoding.jsonEncoder().encode(event2)
+        let decoded = String(data: encoded, encoding: .utf8)
+        let expected = "{\"dimensions\":{\"stop\":\"drop\"}}"
+        XCTAssertEqual(decoded, expected)
     }
 
-    func testCheck() {
-
-        let healthOfServer = "ok"
-        let serverResponse = HealthResponse(status: healthOfServer)
-        let encoded: Data!
-        do {
-            encoded = try ParseCoding.jsonEncoder().encode(serverResponse)
-        } catch {
-            XCTFail("Should encode/decode. Error \(error)")
-            return
-        }
-
-        MockURLProtocol.mockRequests { _ in
-            return MockURLResponse(data: encoded, statusCode: 200, delay: 0.0)
-        }
-        do {
-            let health = try ParseHealth.check()
-            XCTAssertEqual(health, healthOfServer)
-
-        } catch {
-            XCTFail(error.localizedDescription)
-        }
-    }
-
-    func testCheckAsync() {
-        let healthOfServer = "ok"
-        let serverResponse = HealthResponse(status: healthOfServer)
+    #if canImport(UIKit)
+    func testTrackAppOpenedUIKit() {
+        let serverResponse = NoBody()
         let encoded: Data!
         do {
             encoded = try ParseCoding.jsonEncoder().encode(serverResponse)
@@ -89,14 +76,37 @@ class ParseAnalyticsTests: XCTestCase {
             return MockURLResponse(data: encoded, statusCode: 200, delay: 0.0)
         }
 
-        let expectation = XCTestExpectation(description: "Health check")
-        ParseHealth.check { result in
-            switch result {
+        let expectation = XCTestExpectation(description: "Analytics save")
+        let options = [UIApplication.LaunchOptionsKey.remoteNotification: ["stop": "drop"]]
+        ParseAnalytics.trackAppOpened(launchOptions: options) { result in
 
-            case .success(let health):
-                XCTAssertEqual(health, healthOfServer)
+            if case .failure(let error) = result {
+                XCTFail(error.localizedDescription)
+            }
+            expectation.fulfill()
+        }
+        wait(for: [expectation], timeout: 10.0)
+    }
+    #endif
 
-            case .failure(let error):
+    func testTrackAppOpened() {
+        let serverResponse = NoBody()
+        let encoded: Data!
+        do {
+            encoded = try ParseCoding.jsonEncoder().encode(serverResponse)
+        } catch {
+            XCTFail("Should encode/decode. Error \(error)")
+            return
+        }
+
+        MockURLProtocol.mockRequests { _ in
+            return MockURLResponse(data: encoded, statusCode: 200, delay: 0.0)
+        }
+
+        let expectation = XCTestExpectation(description: "Analytics save")
+        ParseAnalytics.trackAppOpened(dimensions: ["stop": "drop"]) { result in
+
+            if case .failure(let error) = result {
                 XCTFail(error.localizedDescription)
             }
             expectation.fulfill()
@@ -104,11 +114,11 @@ class ParseAnalyticsTests: XCTestCase {
         wait(for: [expectation], timeout: 10.0)
     }
 
-    func testCheckErrorAsync() {
-        let healthOfServer = "Should throw error"
+    func testTrackEvent() {
+        let serverResponse = NoBody()
         let encoded: Data!
         do {
-            encoded = try ParseCoding.jsonEncoder().encode(healthOfServer)
+            encoded = try ParseCoding.jsonEncoder().encode(serverResponse)
         } catch {
             XCTFail("Should encode/decode. Error \(error)")
             return
@@ -118,10 +128,12 @@ class ParseAnalyticsTests: XCTestCase {
             return MockURLResponse(data: encoded, statusCode: 200, delay: 0.0)
         }
 
-        let expectation = XCTestExpectation(description: "Health check")
-        ParseHealth.check { result in
-            if case .success = result {
-                XCTFail("Should have thrown error")
+        let expectation = XCTestExpectation(description: "Analytics save")
+        let event = ParseAnalytics(name: "hello")
+        event.track { result in
+
+            if case .failure(let error) = result {
+                XCTFail(error.localizedDescription)
             }
             expectation.fulfill()
         }