+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Parse-Swift Changelog
 
 ### main
-[Full Changelog](https://github.com/parse-community/Parse-Swift/compare/1.7.2...main)
+[Full Changelog](https://github.com/parse-community/Parse-Swift/compare/1.7.3...main)
 * _Contributing to this repo? Add info about your change here to be included in the next release_
+
+### 1.7.3
+[Full Changelog](https://github.com/parse-community/Parse-Swift/compare/1.7.2...1.7.3)
 
 __New features__
 - Add ParseAnalytics  ([#147](https://github.com/parse-community/Parse-Swift/pull/147)), thanks to [Corey Baker](https://github.com/cbaker6).